requirements.txt
torch>=2.6.0
torchvision>=0.21.0
timm>=1.0.15
mxnet>=1.9.1
opencv-python>=4.10.0.84
numpy>=1.26.0,<2.0.0
pytorch-lightning>=2.5.1
tqdm
imgaug
accelerate>=1.6.0
scikit-learn
pillow
zipfile
requests